+++ packages/contracts/src/lib/zapier.model.ts
@@ -62,13 +62,6 @@ export interface IZapierEndpoint {
 	slug: string;
 }
 
-export interface IZapierAccessTokens {
-	access_token: string;
-	refresh_token: string;
-	token_type: string;
-	expires_in: number;
-}
-
 export interface ICreateZapierIntegrationInput extends IBasePerTenantAndOrganizationEntityModel {
 	client_id: string;
 	client_secret: string;