+++ packages/plugins/integration-zapier-ui/CHANGELOG.md
@@ -5,4 +5,4 @@
 ### Added
 
 -   **IntegrationZapierAuthorizeComponent**: Handles authorization processes for Integration Zapier.
--   **IntegrationZapierComponent**: Provides a UI component for Zapier integration settings management
+-   **IntegrationZapierComponent**: Provides a UI component for Zapier integration settings management.